refactor: move hook to hook file
